// apps/web-frontend/src/api/generated/ship/types.gen.ts
export type Ship = {
    user_id: string;
    hull_health: number;
    materials_balance: number;
    level: number;
    /**
     * RFC3339Nano timestamp.
     */
    updated_at: string;
};